More accurate time display
Removed the section of code that increments the time variable since it
seems to do it several times if any buttons are pressed. Instead, if
seconds display is turned on the time is synced from the TCF0 timer
every second as part of the RTC 1 second ISR. This keeps the display
much closer to matching reality then before where pressing buttons would
skew the displayed time many seconds until the next 1 minute ISR fired
to sync the time from TCF0. This is not a perfect solution, but it is
better! Now the time displayed is within 1 second of the time as tracked
by TCF0 and RTC time-keeping.